Split parallel runner orchestration by responsibility
